server/requirements.txt
streamlit
keras 
matplotlib
seaborn
Pillow
tensorflow==2.8.2
pydicom
opencv-python